outputs.js --> TS
