+++ dbmongo/js/reduce.algo2/ava_tests.ts
@@ -8,7 +8,7 @@ import test, { ExecutionContext } from "ava"
 import "../globals"
 import { map } from "./map"
 import { flatten } from "./flatten.js"
-import { outputs } from "./outputs.js"
+import { outputs } from "./outputs"
 import { repeatable } from "./repeatable"
 import { add } from "./add"
 import { defaillances } from "./defaillances.js"
